LLM Projects/Resume Analyser App/requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
docxtpl